feat(less-to-js): #74 move variables script
