+++ src/less/tags.less
@@ -24,6 +24,7 @@
 @tag-padding-horizontal: 6px;
 @tag-padding-vertical: 3px;
 @tag-text-transform: uppercase;
+@tag-martins-var: uppercase;
 
 @tag-rounded-border-radius: 500px;
 @tag-rounded-font-size: 65%;
